+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 <!--next-version-placeholder-->
 
+## v6.5.2 (2022-05-04)
+### Fix
+* Fix email tuples ([#1846](https://github.com/ocadotechnology/codeforlife-portal/issues/1846)) ([`a39aa47`](https://github.com/ocadotechnology/codeforlife-portal/commit/a39aa4768e46df7e46e22bcf16c885ca7d37fe1e))
+
+## v6.5.1 (2022-05-03)
+### Fix
+* Remove lat, long, town, agg data, map ([#1845](https://github.com/ocadotechnology/codeforlife-portal/issues/1845)) ([`caf5e96`](https://github.com/ocadotechnology/codeforlife-portal/commit/caf5e96c3bc8c619f2cda16ab266332d87262ba1))
+
+## v6.5.0 (2022-05-03)
+### Feature
+* Anonymise school ([#1844](https://github.com/ocadotechnology/codeforlife-portal/issues/1844)) ([`0e5eaf5`](https://github.com/ocadotechnology/codeforlife-portal/commit/0e5eaf553999131095d9df4b1b29a8a848ccc687))
+
 ## v6.4.4 (2022-04-28)
 ### Fix
 * Email error ([#1842](https://github.com/ocadotechnology/codeforlife-portal/issues/1842)) ([`9fe37d7`](https://github.com/ocadotechnology/codeforlife-portal/commit/9fe37d7bd5cd81903f57127d0abe7568c885b3b7))